refractor: move css files to styles, added component tree folders
Component tree is now represented in the folder structure and all css files are in the current folder
