+++ src/components/React/controls/OpacityControl.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react'
-import '../../App.css'
+import "../../../styles/App.css"
 import { connect } from "react-redux";
-import {mySendAlphaPoints} from '../../redux/AppActions'
+import {mySendAlphaPoints} from "../../../redux/AppActions"
 
 const mapStateToProps = state => {
   return {volumeData: state.volumeData};
